Fixed lighting. Added a flashlight.
The design is moving towards a more intense experience, like Dead Space or System Shock. I need to limit the ammo player has and finish my pickup system so the player starts the adventure with only a pistol or a revolver - I'll decide which. Then the player has to clear rooms, find pickups, activate switches/buttons and progress through the level. Week 5 is going to be dedicated to this.